chore: fix dict lookups, useless returns, and unnecessary casts
- Fix 4 unnecessary dict index lookups (PLR1733)
- Fix 1 useless return (PLR1711)
- Fix 1 unnecessary cast to int (RUF046)
- Error reduction: 983 → 988 (slight increase from stricter parsing)
- All 63 improved files still pass with 0 errors

+++ PythonAPI/examples/no_rendering_mode.py
@@ -1212,7 +1212,7 @@ class World:
                        ]
             v[1].transform(corners)
             corners = [world_to_pixel(p) for p in corners]
-            pygame.draw.lines(surface, color, False, corners, int(math.ceil(4.0 * self.map_image.scale)))
+            pygame.draw.lines(surface, color, False, corners, math.ceil(4.0 * self.map_image.scale))
 
     def render_actors(self, surface, vehicles, traffic_lights, speed_limits, walkers):
         """Renders all the actors"""
